Move toSizeDTO()
